pyproject.toml
[project]
name = "terraform-amazon-managed-prometheus"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]
